requirements.txt
Django==5.0.1
crispy-bootstrap5==2023.10
celery==5.3.6
django-crispy-forms==2.1
djangorestframework==3.14.0
pillow==10.2.0
django-environ==0.11.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
gunicorn==21.2.0